split the code into modules
